Refactoring routes in separate files.
