+++ routes/index.js
@@ -1,0 +1,58 @@
+var express  = require("express"),
+    router   = express.Router(),
+    passport = require("passport"),
+    User     = require("../models/user");
+
+// "/" Landing Page
+router.get("/", function (req, res) {
+    res.render("landing");
+});
+
+
+//show register form
+router.get("/register", function (req, res) {
+    res.render("register");
+});
+
+// Create a user
+router.post("/register", function (req, res) {
+    var newUser = new User({username: req.body.username});
+    User.register(newUser, req.body.password, function (err, user) {
+        if(err) {
+            console.log(err);
+            return res.render("register");
+        }
+        passport.authenticate("local")(req, res, function () {
+            res.redirect("/holygrounds")
+        });
+    });
+});
+
+// LogIn
+router.get("/login", function (req, res) {
+    res.render("login");
+});
+
+// handling login logic
+router.post("/login", passport.authenticate("local",
+    {
+        successRedirect: "/holygrounds",
+        failureRedirect: "/login"
+    }), function(req, res){
+});
+
+
+router.get("/logout", function (req, res) {
+    req.logout();
+    res.redirect("/holygrounds");
+});
+
+// Middleware todo refactor in a new file
+function isLoggedIn(req, res, next){
+    if(req.isAuthenticated()){
+        return next();
+    }
+    res.redirect("/login");
+}
+
+module.exports = router;
